requirements.txt
RPi.GPIO==0.7.1
pigpio==1.78
